`jquery` package extraction started
