+++ framework/jquery/ActiveField.php
@@ -1,0 +1,22 @@
+<?php
+/**
+ * @link http://www.yiiframework.com/
+ * @copyright Copyright (c) 2008 Yii Software LLC
+ * @license http://www.yiiframework.com/license/
+ */
+
+namespace yii\jquery;
+
+/**
+ * ActiveField represents a form input field within an [[ActiveForm]].
+ *
+ * @see ActiveForm
+ *
+ * @author Qiang Xue <qiang.xue@gmail.com>
+ * @author Paul Klimov <pklimov@quartsoft.com>
+ * @since 2.1
+ */
+class ActiveField extends \yii\widgets\ActiveField
+{
+
+}
