+++ framework/jquery/ActiveFormAsset.php
@@ -5,29 +5,24 @@
  * @license http://www.yiiframework.com/license/
  */
 
-namespace yii\validators;
+namespace yii\jquery;
 
 use yii\web\AssetBundle;
-use yii\web\YiiAsset;
 
 /**
- * This asset bundle provides the javascript files for client validation.
- *
  * @author Qiang Xue <qiang.xue@gmail.com>
  * @since 2.0
  */
-class ValidationAsset extends AssetBundle
+class ActiveFormAsset extends AssetBundle
 {
     /**
      * @inheritdoc
      */
     public $sourcePath = '@yii/assets';
-
     /**
      * @inheritdoc
      */
-    public $js = ['yii.validation.js'];
-
+    public $js = ['yii.activeForm.js'];
     /**
      * @inheritdoc
      */
