+++ framework/jquery/ValidationAsset.php
@@ -5,27 +5,26 @@
  * @license http://www.yiiframework.com/license/
  */
 
-namespace yii\widgets;
+namespace yii\jquery;
 
 use yii\web\AssetBundle;
-use yii\web\YiiAsset;
 
 /**
+ * This asset bundle provides the javascript files for client validation.
+ *
  * @author Qiang Xue <qiang.xue@gmail.com>
  * @since 2.0
  */
-class ActiveFormAsset extends AssetBundle
+class ValidationAsset extends AssetBundle
 {
     /**
      * @inheritdoc
      */
-    public $sourcePath = '@yii/assets';
-
+    public $sourcePath = '@yii/jquery/assets';
     /**
      * @inheritdoc
      */
-    public $js = ['yii.activeForm.js'];
-
+    public $js = ['yii.validation.js'];
     /**
      * @inheritdoc
      */
