+++ framework/jquery/assets/yii.captcha.js
